move sample-Counter
