+++ how-much-android/redux/src/main/java/com/spread/redux/Action.kt
@@ -1,3 +1,3 @@
 package com.spread.redux
 
-sealed interface Action
+interface Action
