+++ how-much-android/redux/src/main/java/com/spread/redux/HowMuchState.kt
@@ -1,6 +1,6 @@
 package com.spread.redux
 
 data class HowMuchState(
-    val counterState: CounterState = CounterState()
+    val counterState: com.spread.redux.sample.CounterState = _root_ide_package_.com.spread.redux.sample.CounterState()
 )
