+++ how-much-android/redux/src/main/java/com/spread/redux/sample/Counter.kt
@@ -1,4 +1,4 @@
-package com.spread.redux
+package com.spread.redux.sample
 
 import org.reduxkotlin.typedReducer
