+++ how-much-android/redux/src/main/java/com/spread/redux/sample/CounterAction.kt
@@ -1,4 +1,6 @@
-package com.spread.redux
+package com.spread.redux.sample
+
+import com.spread.redux.Action
 
 sealed class CounterAction: Action {
     class Increment(val value: Int = 1) : CounterAction()
